api/pyproject.toml
[project]
name = "devpush-api"
version = "0.1.0"
description = "API for devpush"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-cli>=0.0.7",
    "jinja2>=3.1.6",
    "pyjwt>=2.10.1",
    "redis>=6.1.0",
    "uvicorn>=0.34.2",
    "asyncpg>=0.29.0",
    "authlib>=1.6.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "itsdangerous>=2.2.0",
    "cryptography>=45.0.4",
    "sqlalchemy[asyncio]>=2.0.41",
    "alembic>=1.16.2",
    "psycopg2-binary>=2.9.10",
    "humanize>=4.12.3",
    "python-multipart>=0.0.20",
    "starlette-wtf>=0.4.5",
    "arq>=0.25.0",
    "aiodocker>=0.24.0",
    "greenlet>=3.2.3",
    "watchfiles>=1.1.0",
    "pillow>=11.3.0",
    "types-pillow>=10.2.0.20240822",
]
